Renaming qa_utils to utils and sdc_client to sdcclient
