+++ python-sdcclient/sdcclient/productandrelease_resource_client.py
@@ -22,9 +22,9 @@
 # contact with opensource@tid.es
 
 
-from qa_utils.rest_client_utils import RestClient, API_ROOT_URL_ARG_NAME, model_to_request_body, HEADER_CONTENT_TYPE, \
+from utils.rest_client_utils import RestClient, API_ROOT_URL_ARG_NAME, model_to_request_body, HEADER_CONTENT_TYPE, \
     HEADER_REPRESENTATION_XML
-from qa_utils.logger_utils import get_logger
+from utils.logger_utils import get_logger
 
 logger = get_logger("sdcClient")
